pyproject.toml
[project]
name = "langchain-practice"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "black>=25.11.0",
    "isort>=6.1.0",
    "langchain>=0.3.27",
    "langchain-ollama>=0.3.10",
    "langchain-openai>=0.3.35",
    "langsmith>=0.4.37",
    "python-dotenv>=1.2.1",
]
